Fixed a bug in OrderItem Entity, place order DTO was added, OrderService checks for customer

+++ back-end/src/main/java/com/project/code/Model/OrderItem.java
@@ -27,7 +27,7 @@ public class OrderItem {
     private Double price;
 
 
-    public OrderItem(OrderDetails order, Product product, int price){
+    public OrderItem(OrderDetails order, Product product, Double price){
         this.order = order;
         this.product = product;
         this.price = price;
